Tweaks to inbox view - #70 - moved inboxCard from index.tsx to its own file. - path line breaks and icon fixed - should still do something about these buttons :P
